Preliminary (but slow) support for tied-mixture (i.e. genones or PTM) models.  Stil testing to make sure it actually helps.
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/pocketsphinx@9617 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/libpocketsphinx/ms_senone.h
@@ -157,8 +157,8 @@ typedef struct {
     uint32 n_cw;		/**< Number codewords per codebook,stream */
     uint32 n_gauden;		/**< Number gaussian density codebooks referred to by senones */
     float32 mixwfloor;		/**< floor applied to each PDF entry */
-    int16 *mgau;		/**< senone-id -> mgau-id mapping for senones in this set */
-    int32* featscr;              /**< The feature score for every senone, will be initialized inside senone_eval_all */
+    uint32 *mgau;		/**< senone-id -> mgau-id mapping for senones in this set */
+    int32 *featscr;              /**< The feature score for every senone, will be initialized inside senone_eval_all */
 } senone_t;
